+++ medical_clinic_project/src/main/java/com/example/medical_clinic_project/Repository/FisaMedicalaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface FisaMedicalaRepository extends JpaRepository<FisaMedicala, Integer> {
+public interface FisaMedicalaRepository extends JpaRepository<FisaMedicala, Long> {
 }
