+++ medical_clinic_project/src/main/java/com/example/medical_clinic_project/Repository/MedicRepository.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface MedicRepository extends JpaRepository<Medic, Integer> {
+public interface MedicRepository extends JpaRepository<Medic, Long> {
     @Query("SELECT m FROM Medic m WHERE m.cnpMedic = ?1")
     Optional<Medic> findMedicByCnp(String cnp);
 }
