+++ medical_clinic_project/src/main/java/com/example/medical_clinic_project/Repository/PretRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PretRepository extends JpaRepository<Pret, Integer> {
+public interface PretRepository extends JpaRepository<Pret, Long> {
 }
