+++ medical_clinic_project/src/main/java/com/example/medical_clinic_project/Repository/SpecializareRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SpecializareRepository extends JpaRepository<Specializare, Integer> {
+public interface SpecializareRepository extends JpaRepository<Specializare, Long> {
 }
